Use CopyFile rather than CopyDiskToMachineDir
CopyDiskToMachineDir will assume that the URL we gave it is an actual
URL, and will call a method called 'Download'. Since all we want is to
copy the disk image from one location to another, we can directly call
the appropriate method and skip the indirection.
CopyDiskToMachineDir is gone in newer versions of the machine package,
and Driver.DiskPathURL is replaced by Driver.ImageSourcePath, this
commit adjusts to these changes.