Introduce findHyperkitProcess helper, and use it in recoverFromUncleanShutdown
The end goal is to remove getPid(), and to replace it with
findHyperkitProcess(). The main motivation for this is that the newer
hyperkit go bindings don't support deserialization of hyperkit.json,
which we rely on to get the hyperkit pid. This pid is needed by
GetState, recoverFromUncleanShutdown and sendSignals..

Since we we have a hyperkit.pid file containing just the hyperkit pid,
we can make use of it instead of parsing a json file.